# contract/openapi.yaml
openapi: 3.1.0

info:
  title: Tournament Manager API
  version: 1.0.0
  summary: API de gestion de tournoi
  description: |
    API REST pour gérer un tournoi.

    Cette version de l'API se concentre sur trois concepts principaux :
    - les équipes
    - les matchs
    - le classement

    Principes de conception :
    - une équipe ne contient pour l'instant qu'un identifiant et un nom
    - un match référence un ensemble d'équipes participantes
    - un score est toujours simple, sur un seul set
    - le score est manipulé via une sous-ressource dédiée `/games/{gameId}/score`
    - le classement est exposé via `/rankings`

    Cette spécification est destinée à devenir la source de vérité du serveur.

servers:
  - url: "http://localhost:3000/api"
    description: "Serveur local"

tags:
  - name: Teams
    description: |
      Gestion du référentiel des équipes.
      Une équipe est actuellement définie uniquement par son nom.
  - name: Games
    description: |
      Gestion des matchs.
      Un match possède une date planifiée, un terrain, un statut
      et un ensemble d'équipes participantes sans doublon.
  - name: Scores
    description: |
      Gestion du score d'un match.
      Dans cette version, le score est toujours un score simple,
      représenté comme une association entre identifiant d'équipe et nombre de points.
  - name: Rankings
    description: |
      Exposition du classement calculé à partir des matchs et de leurs scores.

paths:
  /teams:
    get:
      tags: [Teams]
      operationId: listTeams
      summary: Lister les équipes
      description: Retourne l'ensemble des équipes connues du système.
      responses:
        '200':
          description: Liste des équipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
              examples:
                default:
                  summary: Exemple de liste d'équipes
                  value:
                    - id: team_1
                      name: Les Furets
                    - id: team_2
                      name: Les Blaireaux

    post:
      tags: [Teams]
      operationId: createTeam
      summary: Créer une équipe
      description: Crée une nouvelle équipe.
      requestBody:
        required: true
        description: Informations de création de l'équipe
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
            examples:
              default:
                summary: Création d'une équipe
                value:
                  name: Les Furets
      responses:
        '201':
          description: Équipe créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
              examples:
                default:
                  value:
                    id: team_1
                    name: Les Furets
        '400':
          $ref: '#/components/responses/BadRequest'

  /teams/{teamId}:
    parameters:
      - $ref: '#/components/parameters/TeamId'
    get:
      tags: [Teams]
      operationId: getTeamById
      summary: Lire une équipe
      description: Retourne une équipe à partir de son identifiant.
      responses:
        '200':
          description: Équipe trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Teams]
      operationId: updateTeam
      summary: Modifier une équipe
      description: Met à jour le nom d'une équipe existante.
      requestBody:
        required: true
        description: Nouveau contenu de l'équipe
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRequest'
            examples:
              default:
                value:
                  name: Les Furets Volants
      responses:
        '200':
          description: Équipe mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Teams]
      operationId: deleteTeam
      summary: Supprimer une équipe
      description: Supprime une équipe existante.
      responses:
        '204':
          description: Équipe supprimée
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            L'équipe ne peut pas être supprimée car elle est encore référencée
            par un ou plusieurs matchs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                teamUsedInGame:
                  value:
                    code: TEAM_IN_USE
                    message: Team team_1 is still referenced by existing games

  /games:
    get:
      tags: [Games]
      operationId: listGames
      summary: Lister les matchs
      description: Retourne l'ensemble des matchs connus du système.
      responses:
        '200':
          description: Liste des matchs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
              examples:
                default:
                  value:
                    - id: game_1
                      scheduledAt: '2026-03-15T18:30:00Z'
                      court: Court A
                      status: scheduled
                      teams:
                        - id: team_1
                          name: Les Furets
                        - id: team_2
                          name: Les Blaireaux
                      score: null

    post:
      tags: [Games]
      operationId: createGame
      summary: Créer un match
      description: |
        Crée un nouveau match.

        Le match référence un ensemble d'équipes participantes, représenté côté contrat
        par un tableau sans doublon (`uniqueItems: true`).

        Dans cette première version :
        - un match contient au moins 2 équipes
        - les équipes doivent déjà exister
        - le score n'est pas fourni ici ; il est géré via `/games/{gameId}/score`
      requestBody:
        required: true
        description: Informations de création du match
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
            examples:
              default:
                value:
                  scheduledAt: '2026-03-15T18:30:00Z'
                  court: Court A
                  teamIds:
                    - team_1
                    - team_2
      responses:
        '201':
          description: Match créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Une ou plusieurs équipes référencées sont introuvables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unknownTeam:
                  value:
                    code: TEAM_NOT_FOUND
                    message: One or more teams do not exist

  /games/{gameId}:
    parameters:
      - $ref: '#/components/parameters/GameId'
    get:
      tags: [Games]
      operationId: getGameById
      summary: Lire un match
      description: Retourne le détail d'un match.
      responses:
        '200':
          description: Match trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Games]
      operationId: updateGame
      summary: Modifier un match
      description: |
        Met à jour complètement les informations principales du match.

        Le score n'est pas modifié par cette opération.
        Pour modifier le score, utiliser `/games/{gameId}/score`.
      requestBody:
        required: true
        description: Nouveau contenu du match
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameRequest'
            examples:
              default:
                value:
                  scheduledAt: '2026-03-15T19:00:00Z'
                  court: Court B
                  status: scheduled
                  teamIds:
                    - team_1
                    - team_2
      responses:
        '200':
          description: Match mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Games]
      operationId: deleteGame
      summary: Supprimer un match
      description: Supprime un match existant ainsi que son score éventuel.
      responses:
        '204':
          description: Match supprimé
        '404':
          $ref: '#/components/responses/NotFound'

  /games/{gameId}/score:
    parameters:
      - $ref: '#/components/parameters/GameId'
    get:
      tags: [Scores]
      operationId: getGameScore
      summary: Lire le score d'un match
      description: |
        Retourne le score associé à un match.

        Le score est un score simple sur un seul set.
        Il associe chaque équipe participante à son nombre de points.
      responses:
        '200':
          description: Score trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameScore'
              examples:
                default:
                  value:
                    pointsByTeam:
                      team_1: 21
                      team_2: 18
        '404':
          description: Match introuvable ou score absent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                gameNotFound:
                  value:
                    code: GAME_NOT_FOUND
                    message: Game game_1 not found
                scoreNotFound:
                  value:
                    code: SCORE_NOT_FOUND
                    message: No score found for game game_1

    put:
      tags: [Scores]
      operationId: upsertGameScore
      summary: Créer ou remplacer le score d'un match
      description: |
        Crée ou remplace intégralement le score d'un match.

        Cette opération permet :
        - de créer un score si le match n'en possède pas encore
        - de remplacer le score existant

        Le payload doit contenir exactement une entrée par équipe participante au match.
      requestBody:
        required: true
        description: Nouveau score du match
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertGameScoreRequest'
            examples:
              default:
                value:
                  pointsByTeam:
                    team_1: 21
                    team_2: 18
      responses:
        '200':
          description: Score enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameScore'
        '400':
          description: Le score fourni est invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidScore:
                  value:
                    code: INVALID_SCORE
                    message: Score must contain exactly one entry for each participating team
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Scores]
      operationId: deleteGameScore
      summary: Supprimer le score d'un match
      description: Supprime le score associé à un match.
      responses:
        '204':
          description: Score supprimé
        '404':
          $ref: '#/components/responses/NotFound'

  /rankings:
    get:
      tags: [Rankings]
      operationId: listRankings
      summary: Lire le classement
      description: |
        Retourne le classement global des équipes.

        Le classement est une vue calculée à partir des matchs et des scores saisis.
        L'ordre retourné doit être celui du classement métier.
      responses:
        '200':
          description: Classement global
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RankingEntry'
              examples:
                default:
                  value:
                    - rank: 1
                      team:
                        id: team_1
                        name: Les Furets
                      played: 3
                      won: 3
                      lost: 0
                      pointsFor: 63
                      pointsAgainst: 42
                      pointsDifference: 21
                    - rank: 2
                      team:
                        id: team_2
                        name: Les Blaireaux
                      played: 3
                      won: 1
                      lost: 2
                      pointsFor: 54
                      pointsAgainst: 60
                      pointsDifference: -6

  /rankings/{teamId}:
    parameters:
      - $ref: '#/components/parameters/TeamId'
    get:
      tags: [Rankings]
      operationId: getRankingByTeamId
      summary: Lire le classement d'une équipe
      description: Retourne l'entrée de classement correspondant à une équipe donnée.
      responses:
        '200':
          description: Entrée de classement trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingEntry'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    TeamId:
      name: teamId
      in: path
      required: true
      description: Identifiant d'une équipe
      schema:
        type: string
        examples: [team_1]

    GameId:
      name: gameId
      in: path
      required: true
      description: Identifiant d'un match
      schema:
        type: string
        examples: [game_1]

  responses:
    BadRequest:
      description: La requête est invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                code: BAD_REQUEST
                message: Request payload is invalid

    NotFound:
      description: La ressource demandée est introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                code: NOT_FOUND
                message: Requested resource was not found

  schemas:
    Team:
      type: object
      description: Représentation publique d'une équipe.
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: Identifiant technique unique de l'équipe
          examples: [team_1]
        name:
          type: string
          description: Nom affiché de l'équipe
          examples: [Les Furets]

    CreateTeamRequest:
      type: object
      description: Payload utilisé pour créer une équipe.
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: Nom de l'équipe à créer
          examples: [Les Furets]

    UpdateTeamRequest:
      type: object
      description: Payload utilisé pour mettre à jour une équipe.
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: Nouveau nom de l'équipe
          examples: [Les Furets Volants]

    GameStatus:
      type: string
      description: |
        Statut fonctionnel du match.

        - `scheduled` : le match est planifié mais n'a pas commencé
        - `in_progress` : le match a commencé et est en cours
        - `completed` : le match est terminé
      enum:
        - scheduled
        - in_progress
        - completed

    GameScore:
      type: object
      description: |
        Score simple d'un match sur un seul set.

        Le score associe chaque identifiant d'équipe participante
        au nombre de points qui lui est attribué.
      required:
        - pointsByTeam
      properties:
        pointsByTeam:
          type: object
          description: |
            Association entre identifiant d'équipe et nombre de points.

            Exemples :
            - `{ "team_1": 21, "team_2": 18 }`
          additionalProperties:
            type: integer
            format: int32
            minimum: 0
          minProperties: 2
          examples:
            - team_1: 21
              team_2: 18

    UpsertGameScoreRequest:
      type: object
      description: Payload utilisé pour créer ou remplacer le score d'un match.
      required:
        - pointsByTeam
      properties:
        pointsByTeam:
          type: object
          description: |
            Association entre identifiant d'équipe et nombre de points.
            Le contenu doit correspondre exactement aux équipes du match.
          additionalProperties:
            type: integer
            format: int32
            minimum: 0
          minProperties: 2
          examples:
            - team_1: 21
              team_2: 18

    Game:
      type: object
      description: |
        Représentation publique d'un match.

        Les équipes participantes sont retournées sous forme enrichie,
        tandis que les opérations d'écriture utilisent des identifiants.
      required:
        - id
        - scheduledAt
        - court
        - status
        - teams
        - score
      properties:
        id:
          type: string
          description: Identifiant technique unique du match
          examples: [game_1]
        time:
          type: string
          format: date-time
          description: Date et heure planifiées du match
          examples:
            - '2026-03-15T18:30:00Z'
        court:
          type: string
          minLength: 1
          description: Nom ou libellé du terrain
          examples: [Court A]
        status:
          $ref: '#/components/schemas/GameStatus'
        teams:
          type: array
          description: |
            Ensemble des équipes participant au match.

            Le contrat représente cet ensemble comme un tableau JSON sans doublon.
          minItems: 2
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Team'
        referee:
          allOf:
            - $ref: '#/components/schemas/Team'
          nullable: true
        score:
          allOf:
            - $ref: '#/components/schemas/GameScore'
          nullable: true
          description: |
            Score du match.

            Ce champ vaut `null` tant qu'aucun score n'a été saisi.

    CreateGameRequest:
      type: object
      description: Payload utilisé pour créer un match.
      required:
        - scheduledAt
        - court
        - teamIds
      properties:
        scheduledAt:
          type: string
          format: date-time
          description: Date et heure prévues du match
          examples:
            - '2026-03-15T18:30:00Z'
        court:
          type: string
          minLength: 1
          description: Nom ou libellé du terrain
          examples: [Court A]
        teamIds:
          type: array
          description: |
            Ensemble des identifiants des équipes participantes.

            Ce champ est modélisé comme un tableau JSON sans doublon pour représenter un set.
          minItems: 2
          uniqueItems: true
          items:
            type: string
            examples: [team_1]

    UpdateGameRequest:
      type: object
      description: |
        Payload utilisé pour modifier un match.

        Cette opération ne permet pas de modifier le score.
      required:
        - scheduledAt
        - court
        - status
        - teamIds
      properties:
        scheduledAt:
          type: string
          format: date-time
          description: Nouvelle date et heure prévues du match
          examples:
            - '2026-03-15T19:00:00Z'
        court:
          type: string
          minLength: 1
          description: Nouveau terrain
          examples: [Court B]
        status:
          $ref: '#/components/schemas/GameStatus'
        teamIds:
          type: array
          description: |
            Nouvel ensemble des identifiants des équipes participantes.

            Ce champ est modélisé comme un tableau JSON sans doublon pour représenter un set.
          minItems: 2
          uniqueItems: true
          items:
            type: string
            examples: [team_1]

    RankingEntry:
      type: object
      description: |
        Entrée de classement pour une équipe donnée.

        Cette vue est calculée à partir des matchs et des scores disponibles.
      required:
        - rank
        - team
        - played
        - won
        - lost
        - pointsFor
        - pointsAgainst
        - pointsDifference
      properties:
        rank:
          type: integer
          format: int32
          minimum: 1
          description: Position de l'équipe dans le classement
          examples: [1]
        team:
          $ref: '#/components/schemas/Team'
        played:
          type: integer
          format: int32
          minimum: 0
          description: Nombre de matchs joués
          examples: [3]
        won:
          type: integer
          format: int32
          minimum: 0
          description: Nombre de matchs gagnés
          examples: [2]
        lost:
          type: integer
          format: int32
          minimum: 0
          description: Nombre de matchs perdus
          examples: [1]
        pointsFor:
          type: integer
          format: int32
          description: Total des points marqués
          examples: [63]
        pointsAgainst:
          type: integer
          format: int32
          description: Total des points encaissés
          examples: [57]
        pointsDifference:
          type: integer
          format: int32
          description: Différence entre points marqués et encaissés
          examples: [6]

    ErrorResponse:
      type: object
      description: Format d'erreur standard de l'API.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Code fonctionnel ou technique de l'erreur
          examples:
            - NOT_FOUND
            - BAD_REQUEST
            - INVALID_SCORE
        message:
          type: string
          description: Message décrivant l'erreur
          examples:
            - Requested resource was not found